move MagickWand to wand/magick.rs and filters to filters.rs
