+++ src/wand/mod.rs
@@ -1,0 +1,3 @@
+mod magick;
+
+pub use self::magick::MagickWand;
